Refactor ChainSignatureDashboard and Dashboard components for improved EVM address and balance management
- Removed local state management for EVM address and chain balances in ChainSignatureDashboard, now relying on props passed from the Dashboard component.
- Implemented address derivation and balance fetching logic directly in the Dashboard component, enhancing data flow and reducing redundancy.
- Updated the ChainSignatureDashboard to accept new props for EVM address, derivation status, error handling, and chain balances, improving component reusability.
- Enhanced user experience by ensuring that address and balance data is consistently managed and displayed across components.
